Delete README.md
This project focuses on microalgae classification. Due to a great variety of microalgae and low-quality images, it is difficult to gain a satisfying accuracy. To tackle the challenge, we propose a multi-CNNs model , named GT-CNN, to classify microalgae. Experiments results with public datasets show that the GT-CNN model get a higher accuracy compared with the state-of-art methods.